fix etcd-io/bbolt#247 move compact logic
- add todo
